% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_biocstickers}
\alias{download_biocstickers}
\title{Download BiocStickers}
\usage{
download_biocstickers(where = ".", overwrite = FALSE)
}
\arguments{
\item{where}{Root directory to download stickers.
The stickers will be downloaded in a subdirectory 'originals' within that directory.
Defaults to current directory.}

\item{overwrite}{If \code{FALSE}, skip stickers already present locally.
If \code{TRUE}, re-download all stickers.}
}
\value{
\code{TRUE} if all stickers were downloaded successfully.
Otherwise, \code{FALSE}.
}
\description{
Download BiocStickers
}
\examples{
if (interactive()) {
  download_biocstickers()
}
}
